services/report_service/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
sqlalchemy>=2.0.9
psycopg2-binary>=2.9.6
requests>=2.28.2
python-dotenv>=1.0.0
httpx>=0.24.0
python-jose>=3.3.0
pandas>=2.0.0
matplotlib>=3.7.1
jinja2>=3.1.2
aiofiles>=23.1.0
apscheduler>=3.10.1
